bundle into components
