+++ utils/post.ts
@@ -1,0 +1,18 @@
+
+const Post = async (data:wallet[],uid:string) => {
+    fetch(`/api/post/${uid}`, {
+      method: 'POST', 
+      mode: 'cors', // no-cors, *cors, same-origin
+      cache: 'force-cache', // *default, no-cache, reload, force-cache, only-if-cached
+      credentials: 'same-origin', // include, *same-origin, omit
+      headers: {
+        'Content-Type': 'application/json'
+        // 'Content-Type': 'application/x-www-form-urlencoded',
+      },
+      redirect: 'follow', // manual, *follow, error
+      referrerPolicy: 'no-referrer', // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
+      body: JSON.stringify(data) // body data type must match "Content-Type" header
+    })
+  }
+
+export default Post;
